:lipstick: Restyle close button
