+++ src/lib/Screen.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import CloseButton from '$lib/CloseButton.svelte'
 	import { slide } from 'svelte/transition'
 	import { quadOut } from 'svelte/easing'
 	import { openScreen } from '$lib/store'
@@ -17,8 +16,12 @@
 			openScreen.set(null)
 		}
 	}
+	const handleKeyDown = ({ key }) => {
+		if (key === 'Escape') openScreen.set(null)
+	}
 </script>
 
+<svelte:window on:keydown={handleKeyDown} />
 <div
 	class="background"
 	bind:this={backgroundElement}
@@ -26,11 +29,19 @@
 	on:mousedown={onBackgroundMousedown}
 	on:mouseup={onBackgroundMouseup}
 >
-	<section>
-		<CloseButton />
-		<h2>{title}</h2>
-		<slot />
-	</section>
+	<button title="Close" on:click={() => openScreen.set(null)}>
+		<svg viewBox="0 0 10 10" xmlns="http://www.w3.org/2000/svg" width="24" height="24">
+			<title>Close</title>
+			<line x1="1" y1="1" x2="9" y2="9" />
+			<line x1="9" y1="1" x2="1" y2="9" />
+		</svg>
+	</button>
+	<div class="column">
+		<section>
+			<h2>{title}</h2>
+			<slot />
+		</section>
+	</div>
 </div>
 
 <style>
@@ -43,12 +54,44 @@
 		overflow: auto;
 		background: var(--secondary-color);
 		display: flex;
+		flex-direction: column;
+		justify-content: flex-start;
+		align-items: center;
+	}
+
+	.column {
+		max-width: 520px;
+		width: 100%;
+		display: flex;
 		align-items: flex-start;
 		justify-content: center;
 	}
 
+	button {
+		margin: 0;
+		border: 0;
+		width: 100%;
+		height: 60px;
+		min-height: 60px;
+		background-color: var(--tertiary-color);
+		display: flex;
+		align-items: center;
+		justify-content: center;
+	}
+
+	svg {
+		margin: 0;
+		stroke-width: 1.5;
+		stroke: var(--text-color);
+		opacity: 0.6;
+		fill: none;
+	}
+
+	button:hover svg {
+		opacity: 1;
+	}
+
 	section {
-		max-width: 520px;
 		flex-grow: 100;
 		position: relative;
 		padding: 1rem 1.5rem;
@@ -58,6 +101,6 @@
 	h2 {
 		font-size: 1.8em;
 		text-align: center;
-		margin: 1.5rem 0;
+		margin: 0.4rem 0 1.5rem;
 	}
 </style>
